simpleV2/requirements-clean.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
PyMySQL==1.1.1
python-dotenv==1.1.0
itsdangerous==2.2.0
email-validator==2.2.0
WTForms==3.1.2
requests==2.31.0
bcrypt==4.3.0
SQLAlchemy==2.0.21
